+++ src/components/Typography/Typography.test.tsx
@@ -9,20 +9,20 @@ describe("Typography", () => {
     'should correctly handle prop $font="%s"',
     async (font, [fontSize, lineHeight, fontWeight, letterSpacing]) => {
       const { getByTestId } = renderWithTheme(
-        <Typography data-testid="test" $font={font as FontVariant} />
+        <Typography data-testid="test" $font={font as FontVariant} />,
       );
 
       expect(getByTestId("test")).toHaveStyle("font-family: Lexend,sans-serif");
       expect(getByTestId("test")).toHaveStyle(
-        `font-size: ${Number((fontSize / 16).toFixed(REM_DP))}rem`
+        `font-size: ${Number((fontSize / 16).toFixed(REM_DP))}rem`,
       );
       expect(getByTestId("test")).toHaveStyle(
-        `line-height:  ${Number((lineHeight / 16).toFixed(REM_DP))}rem`
+        `line-height:  ${Number((lineHeight / 16).toFixed(REM_DP))}rem`,
       );
       expect(getByTestId("test")).toHaveStyle(`font-weight: ${fontWeight}`);
       expect(getByTestId("test")).toHaveStyle(
-        `letter-spacing: ${letterSpacing}`
+        `letter-spacing: ${letterSpacing}`,
       );
-    }
+    },
   );
 });