+++ src/pages/beta/[viewType]/curriculum/index.tsx
@@ -10,10 +10,8 @@ import AppLayout from "@/components/AppLayout";
 import Box from "@/components/Box";
 import Flex from "@/components/Flex";
 import MaxWidth from "@/components/MaxWidth/MaxWidth";
-import { Heading, UL, LI, P } from "@/components/Typography";
-import SubjectPhasePicker, {
-  SubjectPhasePickerData,
-} from "@/components/SubjectPhasePicker/SubjectPhasePicker";
+import { Heading, UL, LI, P, Hr } from "@/components/Typography";
+import { SubjectPhasePickerData } from "@/components/SubjectPhasePicker/SubjectPhasePicker";
 import {
   decorateWithIsr,
   getFallbackBlockingConfig,
@@ -21,6 +19,8 @@ import {
 } from "@/node-lib/isr";
 import { ViewType } from "@/common-lib/urls";
 import curriculumApi2023 from "@/node-lib/curriculum-api-2023";
+import CurriculumLandingHero from "@/components/pages/LandingPages/CurriculumLandingHero";
+import Breadcrumbs from "@/components/Breadcrumbs/Breadcrumbs";
 import Illustration from "@/components/Illustration/Illustration";
 import Cover from "@/components/Cover/Cover";
 import { getSizes } from "@/components/CMSImage/getSizes";
@@ -28,55 +28,48 @@ import OakLink from "@/components/OakLink/OakLink";
 import Typography from "@/components/Typography/Typography";
 import Icon from "@/components/Icon/Icon";
 import CMSClient from "@/node-lib/cms";
-import { BlogPostPreview } from "@/common-lib/cms-types";
-import { serializeDate } from "@/utils/serializeDate";
 import BrushBorders from "@/components/SpriteSheet/BrushSvgs/BrushBorders/BrushBorders";
-// import PostList from "@/components/Posts/PostList/PostList";
-// import usePostList from "@/components/Posts/PostList/usePostList";
-// import PostListItem from "@/components/Posts/PostList/PostListItem/PostListItem";
+import {
+  blogToPostListItem,
+  SerializedBlogPostPreview,
+} from "@/components/pages/BlogIndex.page";
+import { serializeDate } from "@/utils/serializeDate";
+import PostListItem from "@/components/Posts/PostList/PostListItem/PostListItem";
 
 export type CurriculumHomePageProps = {
   subjectPhaseOptions: SubjectPhasePickerData;
-  curriculumBlogs: BlogPostPreview[];
+  posts: SerializedBlogPostPreview[];
 };
 
 const CurriculumHomePage: NextPage<CurriculumHomePageProps> = (props) => {
-  const { subjectPhaseOptions, curriculumBlogs } = props;
-  console.log(curriculumBlogs, "< curriculumBlogs");
-
-  // const curriculumBlogProps = usePostList({
-  //   items: curriculumBlogs,
-  //   withImage: true,
-  // });
+  const { subjectPhaseOptions, posts } = props;
+  const curriculumBlogs = posts.map(blogToPostListItem);
 
   return (
     <AppLayout seoProps={BETA_SEO_PROPS} $background={"grey1"}>
-      <Flex $justifyContent={"center"} $background={"pupilsLightGreen"}>
-        <MaxWidth>
-          <Box $ph={[16, 0]} $pb={[18, 48, 48]}>
-            <Heading
-              $font={["heading-5", "heading-4"]}
-              tag={"h1"}
-              $mt={120}
-              $color={"black"}
-            >
-              Curriculum Resources
-            </Heading>
-            <Heading $font={"heading-light-6"} tag={"h2"} $mv={8}>
-              A collection of high quality resources to support you, whether
-              you're looking for exemplars or help with starting from scratch.
-            </Heading>
+      <Flex $justifyContent={"center"} $background={"mint"}>
+        <MaxWidth $ph={16}>
+          <Box $mt={20}>
+            <Breadcrumbs
+              breadcrumbs={[
+                {
+                  oakLinkProps: { page: "home", viewType: "teachers-2023" },
+                  label: "Home",
+                },
+                {
+                  oakLinkProps: {
+                    page: "curriculum-landing-page",
+                    viewType: "teachers-2023",
+                  },
+                  label: "Curriculum resources",
+                },
+              ]}
+            />
+            <Hr $color={"white"} $mb={0} />
           </Box>
-        </MaxWidth>
-      </Flex>
-      <Flex $background={"white"} $pv={[48]}>
-        <MaxWidth>
-          <Box $ph={[16, 0]} $pb={[24]}>
-            <Heading tag={"h2"} $font={"heading-light-6"} $mb={16}>
-              Oak's Curricula
-            </Heading>
-            <SubjectPhasePicker {...subjectPhaseOptions} />
-          </Box>
+          <Flex $mt={[24, 80]} $mb={[80]}>
+            <CurriculumLandingHero subjectPhaseOptions={subjectPhaseOptions} />
+          </Flex>
         </MaxWidth>
       </Flex>
       <Flex $background={"white"} $justifyContent={"center"}>
@@ -85,11 +78,18 @@ const CurriculumHomePage: NextPage<CurriculumHomePageProps> = (props) => {
             $flexDirection={["column", "row"]}
             $justifyContent={"space-between"}
             $alignItems={"center"}
-            $gap={[24, 120]}
+            $gap={[24, 80, 120]}
             $mb={80}
+            $pl={[16, 0]}
+            $pr={[16, 0]}
+            $mt={80}
           >
             <Box>
-              <Cover $width={480} $height={450} $position={"relative"}>
+              <Cover
+                $width={[320, 340, 480]}
+                $height={[300, 320, 450]}
+                $position={"relative"}
+              >
                 <Illustration
                   noCrop
                   sizes={getSizes([500, 800])}
@@ -100,11 +100,12 @@ const CurriculumHomePage: NextPage<CurriculumHomePageProps> = (props) => {
                 />
               </Cover>
             </Box>
+
             <Box $height={"100%"}>
               <Heading tag="h2" $font={["heading-5", "heading-4"]} $mt={12}>
                 Our approach to curriculum
               </Heading>
-              <P $mv={24} $font={"body-1"}>
+              <P $mv={[16, 24]} $font={"body-1"}>
                 Every school’s approach to curriculum design is different.
                 Whether you are starting from scratch or refreshing your
                 existing curricula, it can be helpful to see examples to inspire
@@ -129,20 +130,26 @@ const CurriculumHomePage: NextPage<CurriculumHomePageProps> = (props) => {
                   $display={"flex"}
                   $alignItems={"center"}
                 >
-                  Read more about our approach to curriculum
+                  Read more about our approach
                   <Icon name={"chevron-right"} />
                 </OakLink>
               </Typography>
             </Box>
           </Flex>
-          <Flex $background={"grey1"} $position={"relative"} $height={812}>
-            <Box $pa={48}>
-              <Heading tag="h3" $font={"heading-4"}>
+          <Flex
+            $background={"grey1"}
+            $position={"relative"}
+            $minHeight={812}
+            $ml={[0, 12, 0]}
+            $mr={[0, 12, 0]}
+          >
+            <Box $pl={[24, 48]} $pr={[24, 48]} $pt={48} $mb={[48, 0]}>
+              <Heading tag="h3" $font={"heading-4"} $mb={24}>
                 Our blogs on curriculum design
               </Heading>
-              {/* {curriculumBlogs.length ? (
+              {curriculumBlogs.length ? (
                 <>
-                  <UL $reset>
+                  <UL $reset data-testid="blog-list">
                     {curriculumBlogs.map((item, i) => (
                       <LI key={`PostList-PostListItem-${i}`}>
                         {i !== 0 && <Hr thickness={4} $mv={32} />}
@@ -156,7 +163,7 @@ const CurriculumHomePage: NextPage<CurriculumHomePageProps> = (props) => {
                   </UL>
                   {<Hr thickness={4} $mt={32} $mb={0} />}
                 </>
-              ) : null} */}
+              ) : null}
             </Box>
             <BrushBorders color="grey1" />
           </Flex>
@@ -234,22 +241,37 @@ export const getStaticProps: GetStaticProps<
     previewMode: true,
   });
 
-  const curriculumBlogs = blogs
+  const serializedPost = blogs
     .filter((blog) => {
-      if (
+      return (
         blog.slug === "how-to-design-a-subject-curriculum" ||
         blog.slug === "how-to-refresh-your-curriculum-using-oak-units" ||
         blog.slug === "how-to-design-a-unit-of-study"
-      ) {
-        return blog;
-      }
+      );
     })
     .map(serializeDate);
+
+  // TEMPORARY (HACKY) REARRANGEMENT FOR BLOGS ORDER
+  const posts = serializedPost.sort((a, b) => {
+    const slugOrder: { [key: string]: number } = {
+      "how-to-design-a-subject-curriculum": 1,
+      "how-to-refresh-your-curriculum-using-oak-units": 2,
+      "how-to-design-a-unit-of-study": 3,
+    };
+    const slugA = a.slug;
+    const slugB = b.slug;
+    const orderA = slugOrder[slugA];
+    const orderB = slugOrder[slugB];
+    if (!orderA || !orderB) {
+      throw new Error("Missing order for blog post");
+    }
+    return orderA - orderB;
+  });
 
   const results: GetStaticPropsResult<CurriculumHomePageProps> = {
     props: {
       subjectPhaseOptions: data,
-      curriculumBlogs: curriculumBlogs,
+      posts,
     },
   };
   const resultsWithIsr = decorateWithIsr(results);